pyproject.toml
[project]
name = "model-based-eggroll"
version = "0.1.0"
description = "Investigating EGGROLL as an optimiser for world model learning and policy training in offline RL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jax>=0.9.1",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mbrl"]
